+++ app/Http/Controllers/Web/OAuthController.php
@@ -112,36 +112,4 @@ class OAuthController extends WebController
             .'&id='.$tokenResult->token->id
             .'&ed='.$tokenResult->token->expires_at);
     }
-
-    public function handleCallbackGoogleDrive(Request $request)
-    {
-        $client = Google::client();
-
-        // authorization code from google.
-        $code = $request->get('code');
-
-        if (empty($code)) {
-            return '';
-        }
-
-        // Exchange authorization code for an access token.
-        $accessToken = $client->fetchAccessTokenWithAuthCode($code);
-        $client->setAccessToken($accessToken);
-        $created = \Carbon\Carbon::createFromTimestamp($accessToken['created']);
-        OauthUserToken::create([
-            'provider' => 'google',
-            'access_token' => $accessToken['access_token'],
-            'refresh_token' => $accessToken['refresh_token'],
-            'expires_at' => $created->addSeconds($accessToken['expires_in']),
-            'scope' => $accessToken['scope'],
-        ]);
-    }
-
-    public function unlinkGoogleDrive()
-    {
-        OauthUserToken::where('user_id', auth()->id())
-            ->where('provider', 'google')
-            ->where('scope', 'https://www.googleapis.com/auth/drive.file')
-            ->delete();
-    }
 }
