+++ routes/web.php
@@ -32,11 +32,6 @@ Route::namespace('Web')->middleware('web-middleware')->group(function () {
     Route::get('/oauth/login/callback', 'OAuthController@handleCallback');
     Route::get('/oauth/login/google', 'OAuthController@redirectToGoogle');
     Route::get('/oauth/login/google/callback', 'OAuthController@handleGoogleCallback');
-
-    Route::get('/oauth/login/google/drive', 'OAuthController@requestGoogleDrive');
-    Route::get('oauth/login/google/drive/callback', 'OAuthController@handleCallbackGoogleDrive');
-    Route::delete('oauth/login/google/drive', 'OAuthController@unlinkGoogleDrive');
-
 });
 
 Route::namespace('Web')->middleware(['web-middleware', 'auth:web'])->group(function () {
